test: Move BGP LoadBalancer tests to separate suite
The suite is testing mostly the BGP integration in the agent, and the LB
functionality is already tested by k8sT/Services.go. In addition, we
will see more BGP tests with the ongoing BGP work. So it makes sense to
move it to the separate suite.
