+++ test/k8sT/Services.go
@@ -7,7 +7,6 @@ import (
 	"context"
 	"fmt"
 	"net"
-	"os"
 	"strings"
 	"sync"
 	"time"
@@ -35,8 +34,6 @@ const (
 
 	testDSServiceIPv4 = "testds-service"
 	testDSServiceIPv6 = "testds-service-ipv6"
-
-	lbSvcName = "test-lb-with-ip"
 )
 
 // The 5.4 CI job is intended to catch BPF complexity regressions and as such
@@ -733,14 +730,11 @@ Secondary Interface %s :: IPv4: (%s, %s), IPv6: (%s, %s)`, helpers.DualStackSupp
 				})
 
 				Context("Tests with direct routing", func() {
-
-					var directRoutingOpts = map[string]string{
-						"tunnel":               "disabled",
-						"autoDirectNodeRoutes": "true",
-					}
-
 					BeforeAll(func() {
-						DeployCiliumOptionsAndDNS(kubectl, ciliumFilename, directRoutingOpts)
+						DeployCiliumOptionsAndDNS(kubectl, ciliumFilename, map[string]string{
+							"tunnel":               "disabled",
+							"autoDirectNodeRoutes": "true",
+						})
 					})
 
 					It("Tests NodePort", func() {
@@ -880,125 +874,6 @@ Secondary Interface %s :: IPv4: (%s, %s), IPv6: (%s, %s)`, helpers.DualStackSupp
 						testCurlFromOutsideWithLocalPort(kubectl, ni, svc2URL, 1, false, 64002)
 					})
 
-					SkipContextIf(helpers.DoesNotExistNodeWithoutCilium, "Tests LoadBalancer", func() {
-						var (
-							frr      string // BGP router
-							routerIP string
-
-							bgpConfigMap string
-
-							lbSVC string
-
-							ciliumPodK8s1, ciliumPodK8s2 string
-							testStartTime                time.Time
-						)
-
-						BeforeAll(func() {
-							frr = applyFRRTemplate(kubectl, ni)
-							kubectl.ApplyDefault(frr).ExpectSuccess("Unable to apply rendered tempplate %s", frr)
-
-							Eventually(func() string {
-								frrPod, err := kubectl.GetPodsIPs(helpers.KubeSystemNamespace, "app=frr")
-								if _, ok := frrPod["frr"]; err != nil || !ok {
-									return ""
-								}
-								routerIP = frrPod["frr"]
-								return routerIP
-							}, 30*time.Second, 1*time.Second).Should(Not(BeEmpty()), "BGP router is not ready")
-
-							bgpConfigMap = applyBGPCMTemplate(kubectl, routerIP)
-							kubectl.ApplyDefault(bgpConfigMap).ExpectSuccess("Unable to apply BGP ConfigMap %s", bgpConfigMap)
-
-							RedeployCiliumWithMerge(kubectl, ciliumFilename, directRoutingOpts,
-								map[string]string{
-									"bgp.enabled":                 "true",
-									"bgp.announce.loadbalancerIP": "true",
-
-									"debug.verbose": "datapath", // https://github.com/cilium/cilium/issues/16399
-								})
-
-							lbSVC = helpers.ManifestGet(kubectl.BasePath(), "test_lb_with_ip.yaml")
-							kubectl.ApplyDefault(lbSVC).ExpectSuccess("Unable to apply %s", lbSVC)
-
-							var err error
-							ciliumPodK8s1, err = kubectl.GetCiliumPodOnNode(helpers.K8s1)
-							ExpectWithOffset(1, err).ShouldNot(HaveOccurred(), "Cannot determine cilium pod name")
-							ciliumPodK8s2, err = kubectl.GetCiliumPodOnNode(helpers.K8s2)
-							ExpectWithOffset(1, err).ShouldNot(HaveOccurred(), "Cannot determine cilium pod name")
-							testStartTime = time.Now()
-						})
-
-						AfterAll(func() {
-							kubectl.Delete(frr)
-							kubectl.Delete(bgpConfigMap)
-							kubectl.Delete(lbSVC)
-							// Delete temp files
-							os.Remove(frr)
-							os.Remove(bgpConfigMap)
-						})
-
-						AfterFailed(func() {
-							res := kubectl.CiliumExecContext(
-								context.TODO(),
-								ciliumPodK8s1,
-								fmt.Sprintf(
-									"hubble observe debug-events --since %v -o json",
-									testStartTime.Format(time.RFC3339),
-								),
-							)
-							helpers.WriteToReportFile(
-								res.CombineOutput().Bytes(),
-								"tests-loadbalancer-hubble-observe-debug-events-k8s1.log",
-							)
-							res = kubectl.CiliumExecContext(
-								context.TODO(),
-								ciliumPodK8s2,
-								fmt.Sprintf(
-									"hubble observe debug-events --since %v -o json",
-									testStartTime.Format(time.RFC3339),
-								),
-							)
-							helpers.WriteToReportFile(
-								res.CombineOutput().Bytes(),
-								"tests-loadbalancer-hubble-observe-debug-events-k8s2.log",
-							)
-						})
-
-						It("Connectivity to endpoint via LB", func() {
-							By("Waiting until the Operator has assigned the LB IP")
-							lbIP, err := kubectl.GetLoadBalancerIP(
-								helpers.DefaultNamespace, lbSvcName, 30*time.Second)
-							Expect(err).Should(BeNil(), "Cannot retrieve LB IP for test-lb")
-
-							By("Waiting until the Agents have announced the LB IP via BGP")
-							Eventually(func() string {
-								return kubectl.ExecInHostNetNS(
-									context.TODO(),
-									ni.outsideNodeName,
-									"ip route",
-								).GetStdOut().String()
-							}, 30*time.Second, 1*time.Second).Should(ContainSubstring(lbIP),
-								"BGP router does not have route for LB IP")
-
-							// Check connectivity from outside
-							url := "http://" + lbIP
-							testCurlFromOutside(kubectl, ni, url, 10, false)
-
-							// Patch service to add a LB source range to disallow requests
-							// from the outsideNode
-							kubectl.Patch(helpers.DefaultNamespace, "service", lbSvcName,
-								`{"spec": {"loadBalancerSourceRanges": ["1.1.1.0/24"]}}`)
-							time.Sleep(5 * time.Second)
-							testCurlFailFromOutside(kubectl, ni, url, 1)
-							// Patch again, but this time add outsideNode IP addr
-							kubectl.Patch(helpers.DefaultNamespace, "service", lbSvcName,
-								fmt.Sprintf(
-									`{"spec": {"loadBalancerSourceRanges": ["1.1.1.0/24", "%s/32"]}}`,
-									ni.outsideIP))
-							time.Sleep(5 * time.Second)
-							testCurlFromOutside(kubectl, ni, url, 10, false)
-						})
-					})
 				})
 
 				SkipItIf(func() bool {
