--- test/k8sT/manifests/test_lb_with_ip.yaml
@@ -1,13 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: test-lb-with-ip
-spec:
-  type: LoadBalancer
-  ports:
-  - port: 80
-    targetPort: 80
-    protocol: TCP
-    name: http
-  selector:
-    zgroup: testDS
